requirements.txt
aiocache==0.11.1
aiocqhttp==1.4.0
aiofiles==0.7.0
aiohttp==3.7.4.post0
alive-progress==1.6.2
altgraph==0.16.1
anyio==3.2.1
appdirs==1.4.3
APScheduler==3.6.0
arrow==1.1.1
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.2.0
beautifulsoup4==4.9.3
bilibili-api==4.1.0
binaryornot==0.4.4
blinker==1.4
certifi==2019.3.9
chardet==3.0.4
click==7.1.2
colorama==0.4.4
common==0.1.2
comtypes==1.1.7
cookiecutter==1.7.3
crypto==1.4.1
cssutils==2.2.0
cycler==0.10.0
decorator==4.4.2
Django==3.1.6
dukpy==0.2.2
echarts-china-cities-pypkg==0.0.9
echarts-china-misc-pypkg==0.0.1
echarts-china-provinces-pypkg==0.0.3
echarts-countries-pypkg==0.1.6
echarts-united-kingdom-pypkg==0.0.1
ExifRead==2.1.2
face-recognition-models==0.3.0
fastapi==0.63.0
future==0.17.1
get==2019.4.13
h11==0.9.0
h2==4.0.0
hpack==4.0.0
httpcore==0.12.3
httpx==0.16.1
Hypercorn==0.11.2
hyperframe==6.0.1
idna==2.8
image==1.5.33
imageio==2.9.0
imageio-ffmpeg==0.4.4
importlib-metadata==2.0.0
iso8601==0.1.12
itchat==1.3.10
itsdangerous==2.0.1
javascripthon==0.10
jieba==0.39
Jinja2==2.10.1
jinja2-time==0.2.0
JsonForm==0.0.2
jsonschema==3.2.0
JsonSir==0.0.2
jupyter-echarts-pypkg==0.1.2
kiwisolver==1.0.1
llvmlite==0.36.0
lml==0.0.2
loguru==0.5.3
macholib==1.11
macropy3==1.1.0b2
Mako==1.1.4
MarkupSafe==1.1.1
matplotlib==3.0.3
mcpi==1.2.0
minecraftstuff==1.0.1
moviepy==1.0.3
multidict==5.1.0
Naked==0.1.31
nb-cli==0.3.2
nonebot==1.8.3
nonebot2==2.0.0a10
numba==0.53.1
numpy==1.16.2
opencv-contrib-python==4.1.0.25
opencv-python==4.1.0.25
packaging==20.9
panda3d==1.10.7
pefile==2018.8.8
pexpect==4.8.0
Pillow==6.0.0
post==2019.4.13
poyo==0.5.0
priority==2.0.0
proglog==0.1.9
prompt-toolkit==1.0.14
psutil==5.6.2
ptyprocess==0.7.0
public==2019.4.13
PyAutoGUI==0.9.42
pycairo==1.20.0
pydantic==1.8.2
pyecharts==0.5.11
pyecharts-javascripthon==0.0.6
pyecharts-jupyter-installer==0.0.3
pyfiglet==0.8.post1
pygame==1.9.5
PyGetWindow==0.0.5
pyglet==1.5.7
Pygments==2.9.0
pygtrie==2.4.2
PyInquirer==1.0.3
PyInstaller==3.4
PyMsgBox==1.0.6
PyOpenGL==3.1.0
PyOpenGL-accelerate @ file:///C:/Users/lenovo/Downloads/PyOpenGL_accelerate-3.1.5-cp37-cp37m-win_amd64.whl
pyparsing==2.3.1
pypng==0.0.19
PyPrind==2.11.3
PyQRCode==1.2.1
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PyQtWebEngine==5.15.4
PyQtWebEngine-Qt5==5.15.2
pyreadline==2.1
PyRect==0.1.4
pyrsistent==0.17.3
PyScreeze==0.1.21
python-dateutil==2.8.0
python-dotenv==0.18.0
Python-EasyConfig==0.1.7
python-for-android==0.7.0
python-resources==0.3
python-slugify==5.0.2
pytools==2021.1
PyTweening==1.0.3
pytz==2019.1
PyWavefront==1.3.3
pywifi==1.1.12
pywin32==224
pywin32-ctypes==0.2.0
PyYAML==5.1
qqbot==2.3.11
qrcode==7.2
Quart==0.14.1
query-string==2019.4.13
regex==2021.7.6
request==2019.4.13
requests==2.25.1
Resource==0.2.1
rfc3986==1.5.0
selenium==3.141.0
serial==0.0.97
sh==1.12.14
shellescape==3.4.1
Sift==5.0.1
sip==6.1.0
six==1.12.0
sniffio==1.2.0
soupsieve==2.2.1
sqlparse==0.4.1
starlette==0.13.6
text-unidecode==1.3
tk==0.1.0
toml==0.10.2
torch==1.8.1
tqdm==4.62.0
typing-extensions==3.10.0.0
tzlocal==1.5.1
urllib3==1.24.1
uvicorn==0.11.8
warp==0.0.1
wcwidth==0.2.5
websockets==8.1
Werkzeug==2.0.1
win32-setctime==1.0.3
wsproto==1.0.0
yarl==1.6.3
zipp==3.3.1
